+++ src/main/java/com/example/kafkatest/configuration/streams/KafkaStreamsConfig.java
@@ -1,4 +1,4 @@
-package com.example.kafkatest.configuration;
+package com.example.kafkatest.configuration.streams;
 
 import avro.articles.TrendingArticles;
 import com.example.ProblemSolving;
